facets/requirements.txt
django
djangorestframework
fonttools
kiwisolver
numpy
packaging
pyparsing
python-dateutil
pytz
six
sqlparse
music21
drf-yasg